添加背包装备功能，修复商店购买bug

+++ Assets/Scripts/Data/ItemDatas.cs
@@ -18,7 +18,7 @@ public class ItemData
     public Sprite sprite;
     public string name;
     public int incraseHp;
-    public float attack;
-    public float defence;
+    public int attack;
+    public int defence;
 
 }
